# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "API untuk aplikasi Job Portal - Pertemuan 6 PPW2",
        "contact": {
            "name": "Akmal Manggala Putra",
            "email": "akmalmanggala2810@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/dashboard/user": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get user dashboard data",
                "description": "Mendapatkan data dashboard untuk user (latest jobs)",
                "operationId": "7ffd7f51129622474687ad8f874dc196",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "latest_jobs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Software Engineer"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Tech Corp"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Jakarta"
                                                    },
                                                    "salary": {
                                                        "type": "string",
                                                        "example": "10000000-15000000"
                                                    },
                                                    "job_type": {
                                                        "type": "string",
                                                        "example": "Full Time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/admin": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get admin dashboard statistics (Admin only)",
                "description": "Mendapatkan statistik dashboard untuk admin",
                "operationId": "a5f2aff883a246008b38698f93b40800",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statistics": {
                                            "properties": {
                                                "total_jobs": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "active_jobs": {
                                                    "type": "integer",
                                                    "example": 35
                                                },
                                                "total_users": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "total_applications": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "pending_applications": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "latest_jobs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Software Engineer"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Tech Corp"
                                                    },
                                                    "applications_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}